added boss level

+++ SpaceInvaders/src/main/java/com/almasb/fxglgames/spaceinvaders/Config.java
@@ -51,7 +51,7 @@ public final class Config {
     public static final int ENEMY_ROWS = 4;
     public static final int ENEMIES_PER_LEVEL = ENEMIES_PER_ROW * ENEMY_ROWS;
 
-    public static final double BONUS_SPAWN_CHANCE = 0.25;
+    public static final float BONUS_SPAWN_CHANCE = 0.25f;
     public static final int BONUS_MOVE_SPEED = 150;
 
     public static final double PLAYER_MOVE_SPEED = 300;
